fix: Lesson Student Service #1

+++ TPC-UPC/Services/LessonStudentService.cs
@@ -15,6 +15,7 @@ namespace TPC_UPC.Services
         private readonly ILessonStudentRepository _lessonStudentRepository;
         private readonly IUnitOfWork _unitOfWork;
         private readonly ILessonRepository _lessonRepository;
+        private readonly INotificationUserRepository _notificationUserRepository;
         private IStudentRepository _studentRepository;
 
         public LessonStudentService(ILessonStudentRepository lessonStudentRepository, IUnitOfWork unitOfWork, ILessonRepository lessonRepository, IStudentRepository studentRepository)
@@ -32,9 +33,26 @@ namespace TPC_UPC.Services
             if (existingLessonStudent == null)
                 return new LessonStudentResponse("LessonStudent not found");
 
+            //Business rule #24
+            DateTime fechaActual = DateTime.Now;
+            DateTime startDate = existingLessonStudent.Lesson.StartDate;
+
+            if (fechaActual.AddMinutes(15) >= startDate)
+                return new LessonStudentResponse("The time to cancel the reservation has expired");
+               
             try
             {
                 _lessonStudentRepository.Remove(existingLessonStudent);
+
+                //if (lesson.Result.LessonTypeId == 2)
+                //{
+                //    NotificationUser nu = new NotificationUser();
+                //    nu.NotificationId = 904;
+                //    nu.UserId = lesson.Result.TutorId;
+
+                //    await _notificationUserRepository.AddAsync(nu);
+                //}
+
                 await _unitOfWork.CompleteAsync();
 
                 return new LessonStudentResponse(existingLessonStudent);
@@ -102,7 +120,6 @@ namespace TPC_UPC.Services
             }
         }
 
-
         public async Task<LessonStudentResponse> UpdateAsync(int lessonId, int studentId, LessonStudent lessonStudent)
         {
             var existingLessonStudent = await _lessonStudentRepository.FindById(lessonId, studentId);
@@ -125,6 +142,7 @@ namespace TPC_UPC.Services
             }
         }
 
+        //Business rule #3
         public async Task<LessonStudentResponse> SaveFeedbackAsync(int lessonId, int studentId, LessonStudent lessonStudent)
         {
             var existingLessonStudent = await _lessonStudentRepository.FindById(lessonId, studentId);
